Add Scaling popup to node edit dialog, implement and use Scaling class.
The "Scaling" class is initialized with the chosen scaling ("general" or
"schleifenbauer", and use_scaling is set to method _scaling_general or
_scaling_schleifenbauer. To add another scaling type, only a method
named for the new scaling needs to be added, and it will be available
for use (i.e. after being added to the edit node dialog). If a scaling
method needs any initialization, an additional "_init_scaling" method
may be implemented (like "_init_scaling_schleifenbauer").

Scaling is applied for any value which is not of byte_type STR.