Handle single-OID queries in snmpGet method
For "SNMPv1 NoSuchName", use the regular snmpGet method with a block
size of 1. Modified handling of NoSuchName error when blockSize == 1.

Keep track of the number of OIDs results have been returned for, the
final length of msg.varbinds is less than oids.length if some OIDs
can't be found. Call processVarbinds when the number of results matches
the number of OIDs, regardless of whether this particular (final) query
yielded data.

Removed snmpGetIndividualOids method.