Create SNMPv3 sessions when v3 is chosen
sessionKey modified to use community or username, depending of the SNMP
version used. Options is defined before create[V3]Session, for use with
both v3 and v2c/v1 sessions. Context, if configured, is added to options
for v3 sessions.

New method createV3UserObject creates a user object to the spec of the
net-snmp module.