Fix bug where SNMPv3 options shown for v1/v2 node
When the node edit dialog was opened for a new node, or a node
configured with v1 or v2c, some v3 options would be shown. Switch the
order of "change" handler triggering to fix the issue. (Previously, the
"version" handler would first hide all the v3 options, then the
"security_level" handler would show the appropriate fields for the
selected/default security level.)